+++ after.json
@@ -1,0 +1,5 @@
+{
+  "host": "hexlet.io",
+  "timeout": 50,
+  "proxy": "123.234.53.22"
+}
