+++ src/identifier.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace GenDiff\MakeDiffAst;
+namespace GenDiff\Identifier;
 
 use function Funct\Collection\union;
 
-function makeDiffAst($contentBeforeChange, $contentAfterChange)
+function identifier($contentBeforeChange, $contentAfterChange)
 {
     $allKeys = union(array_keys($contentBeforeChange), array_keys($contentAfterChange));
     $makeDiff = function ($acc, $key) use ($contentBeforeChange, $contentAfterChange) {
@@ -14,28 +14,21 @@ function makeDiffAst($contentBeforeChange, $contentAfterChange)
                         "key" => $key,
                         "before" => $contentBeforeChange[$key]
                       ];
-
-            return $acc;
         } elseif (array_key_exists($key, $contentAfterChange) && !array_key_exists($key, $contentBeforeChange)) {
             $acc[] = [
                         "type" => "added",
                         "key" => $key,
                         "after" => $contentAfterChange[$key]
                       ];
-
-            return $acc;
         } elseif (array_key_exists($key, $contentAfterChange) && array_key_exists($key, $contentBeforeChange)) {
             $before = $contentBeforeChange[$key];
             $after = $contentAfterChange[$key];
-
             if (is_array($before) && is_array($after)) {
                 $acc[] = [
                             "type" => "nested",
                             "key" => $key,
-                            "after" => makeDiffAst($before, $after)
+                            "after" => identifier($before, $after)
                           ];
-
-                return $acc;
             } else {
                 $type = ($before === $after) ? "unchanged" : "changed";
                 $acc[] = [
@@ -44,13 +37,10 @@ function makeDiffAst($contentBeforeChange, $contentAfterChange)
                             "before" => $before,
                             "after" => $after
                           ];
-
-                return $acc;
             }
         }
+        return $acc;
     };
-
     $diffAst = array_reduce($allKeys, $makeDiff, []);
-    
     return $diffAst;
 }
