--- src/selectRender.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace GenDiff\SelectRender;
-
-function selectRender($ast, $format)
-{
-    $renderPath = "\GenDiff\Renderers\\$format\\render";
-    
-    return $renderPath($ast);
-}
