+++ src/selecter.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace GenDiff\Selecter;
+
+function selecter($ast, $format)
+{
+    $renderPath = "\GenDiff\Renderers\\$format\\Render";
+
+    return $renderPath($ast);
+}
